4.21

+++ 4/blogilista/controllers/blogs.js
@@ -35,6 +35,17 @@ blogsRouter.post('/', async (request, response) => {
   })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(403).json({ error: 'permission denied' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   })
